feat: Implementing BaseTab as parent class for QueryEditorTab and DocumentTab (#2336)
